RunIUG BossPattern & Stage3-n update
1.
Boss pattern increased from 1~2 to 1~4.
Boss Pattern
1) Falling Lava

2) Fist that flies from the left or right.

3) Lightning ball moving like rubber ball.

4) Columns of fire rising from the floor.

2.
Created stage ground 3-n.
Theme is Rock & Lava.
Added element is lava that is struck on the map itself,
and block of 64x64 size.

3.
update character shooting animation.
When character was attacked, the features shown according to the nature of the object were divided into details.

4. Stage-specific Sprite of obstacles added in previous updates has been added to suit the theme.