Run IUG added 3-n stage related object & Bug fix
1. Added 3-n stage theme resource
Added a general resource mapping images of 3-n themes so that blind stages can run on stage with a variety of themes.

2. Added a stagnant lava Obstacle.
Appears only for stage 3n and modified it to consistently receive fixed damage, not die immediately after the attack.
It can be added through the tile map editor to increase scalability.

3. The bomb was classified as an object with two properties.
If the dropping bomb item and throwing bomb item are pushed to the same item list and managed, there is an error in the existing bomb explosion logic, so I fixed this part.

4. I corrected the wrong printing of the bomb item selection button in the item store.

5. Boss, the third pattern, the Lightning ball pattern, has been modified to a more threatening pattern.

6. Boss's core crystal was modified to shake the screen when attacked.

7. map configuration of the Shelter Stage has been modified.